feat: add reset-local recipe
## Description

In this PR, I attempt to add a `just` recipe to that resets Sprout locally. This includes removing `__pycache__` folders, removing the db `db.sqlite3`, and all migrations. 

Im not too familiar with the commands I have used to achieve this, but the local testing of these seemed to work as expected.