ci: :construction_worker: add coverage workflow (#216)
## Description

Adds workflow to push to coverage report and badge to a coverage branch.
Also moves the linting into the test workflow that I then renamed to
simply "build". Also converted the workflows to use justfile (which I
had to install with a workflow), so that we are standardized at least
onto the build steps in the justfile.

<!-- Please delete as appropriate: -->
Doesn't need a review.