ci: :lock: add GitHub App that makes short-lived tokens (#224)
## Description

Finally figured out how to do this. I set up a GitHub App that will be
able to use a specific token to only push changes to main for updating
versions, rather than my current setup that uses a similar token but
that lasts longer. This token only is used in the workflow then gets
deleted. So improves security.

Closes #196

<!-- Please delete as appropriate: -->
Doesn't need a review. but happy to explain it more.