ci: 👩‍💻 add reusable `release-package` workflow (#247)
# Description

Based on the feedback, I've updated this to use `uv` and to all be
included in a `release-package` workflow instead of separate workflows
for releasing tags and updating the changelog + building distributions
and publishing to TestPyPI and PyPI.

Related to seedcase-project/seedcase-sprout#1404 (to close it we'll need
a workflow in `seedcase-sprout` as well).

This PR needs an in-depth review.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Luke W. Johnston <lwjohnst86@users.noreply.github.com>