feat: :construction_worker: `concurrency` to cancel previous running workflows (#303)
# Description

Workflows like releasing and building the website should stop if
another, newer workflow is started. This prevents errors from running
e.g. several release workflows in a row (which may make the same tag).
Building the website should use the latest commit and cancel previous
ones.

I don't actually know if this will work, or if it needs to be in the
calling workflow (not here).

Closes #275, closes #302

Needs a quick review.